Bunge.net request successful (#6)
* Multi commits (#4)

* Debug: testing embeds

* Fix: .gitignore

* Feat: doing more testing with embedding

Also added KEY to gitignore

* Fix: renamed main.html to index.html

* Fix: Readme was dumb and cringe

* Feat: introduced support for requesting news

I figured out how to make an Oauth2 request to get the webpage of the
latest twab. So far, the only integration is actually sending the
request, but placing it in the extension will come in a future update.

* WIP: Fixing nodejs modules and fetch

* Feat(ts): introduced output to `main.ts`

* Feat(bungie.net) requests to twab successful

The bungie request code is now working properly, and returns the most recent twab or twid. Now I will work on integrating bungie's twitter into the code.